Refactor to fetch data from Jupiter AIP API #192
As Jupiter moves away from Fedora PMPY needs to be reworked in order to
fetch the data directly from Jupiter's new AIP API.

This work in progress as we need to reevaluate the predicates that will
be required for the AIP file. The file formats remain the same for all
received data.

This work includes grabbing and saving the information for:

- object_metadata.n3
- file_order.xml
- file_set_metadata.n3
- original_file_metadata.n3
- content_fixity_report.n3
- The entity's files are copied from the Gluster FS