Change directory layout for community/collections
Community and collection AIPs are created successfully. This change will
remove the associations with files for the community and collection AIPs
as they will never be associated with them.

We changed the VHS cassettes to reflect the changes in the jupiter dev
hostname and added simple tests for the structure of community and
collection AIPs.