WIP Add more preservation events to logs
This work in progress adds more information about all logging events
in order to get a better idea of the behavior on PMPY.

Right now the preservation event loggers are initialized on startup
instead of being instantiated every time there is a preservation event.

All is done in the logger reader method. We could change it so the
preservation loggers are instantiated on their own access.

There are 3 placeholder tests that need to be completed.