Use File.open with a block for simple exception handling and ensuring files are closed
With no associated block, `File.open` is a synonym for `File.new`.  The `new` method returns a `File` instance that you can then read or write from. The file remains open until `close` is called on the instance. Once the file variable goes out of scope, then garbage collection will eventually close it, but this may not happen for a while. Meanwhile, resources are being held open.